--- crates/wasi-common/src/hostcalls/mod.rs
@@ -1,7 +1,0 @@
-mod fs;
-mod misc;
-mod sock;
-
-pub use self::fs::*;
-pub use self::misc::*;
-pub use self::sock::*;
