+++ crates/wasi-common/src/hostcalls_impl/mod.rs
@@ -1,7 +1,9 @@
 mod fs;
 mod fs_helpers;
 mod misc;
+mod sock;
 
 pub(crate) use self::fs::*;
 pub(crate) use self::fs_helpers::PathGet;
 pub(crate) use self::misc::*;
+pub(crate) use self::sock::*;
